codeverse/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
aiohttp==3.9.1
pydantic==2.5.3
typing-extensions==4.9.0
